build(initial-setup): refactor packages names
