+++ application/build.gradle
@@ -3,7 +3,7 @@ plugins {
     id 'java-conventions'
 }
 
-group = 'com.kaua.template.application'
+group = 'com.kaua.order.application'
 
 dependencies {
     implementation(project(':domain'))
