+++ domain/src/main/java/com/kaua/order/domain/Identifier.java
@@ -1,0 +1,6 @@
+package com.kaua.order.domain;
+
+public abstract class Identifier {
+
+    public abstract String getValue();
+}
