+++ domain/src/main/java/com/kaua/order/domain/events/DomainEventPublisher.java
@@ -1,0 +1,7 @@
+package com.kaua.order.domain.events;
+
+@FunctionalInterface
+public interface DomainEventPublisher {
+
+    <T extends DomainEvent> void publish(T event);
+}
