+++ domain/src/main/java/com/kaua/order/domain/exceptions/NoStackTraceException.java
@@ -1,0 +1,12 @@
+package com.kaua.order.domain.exceptions;
+
+public class NoStackTraceException extends RuntimeException {
+
+    public NoStackTraceException(final String message) {
+        this(message, null);
+    }
+
+    public NoStackTraceException(final String message, final Throwable cause) {
+        super(message, cause, true, false);
+    }
+}
