+++ domain/src/main/java/com/kaua/order/domain/utils/IdUtils.java
@@ -1,0 +1,20 @@
+package com.kaua.order.domain.utils;
+
+import java.util.UUID;
+
+public final class IdUtils {
+
+    private IdUtils() {}
+
+    private static UUID generate() {
+        return UUID.randomUUID();
+    }
+    
+    public static String generateIdWithHyphen() {
+        return generate().toString();
+    }
+
+    public static String generateIdWithoutHyphen() {
+        return generate().toString().replace("-", "");
+    }
+}
