+++ domain/src/main/java/com/kaua/order/domain/validation/Error.java
@@ -1,0 +1,4 @@
+package com.kaua.order.domain.validation;
+
+public record Error(String message) {
+}
