+++ domain/src/main/java/com/kaua/order/domain/validation/Validator.java
@@ -1,0 +1,16 @@
+package com.kaua.order.domain.validation;
+
+public abstract class Validator {
+
+    private final ValidationHandler handler;
+
+    protected Validator(final ValidationHandler handler) {
+        this.handler = handler;
+    }
+
+    public abstract void validate();
+
+    protected ValidationHandler validationHandler() {
+        return this.handler;
+    }
+}
