--- domain/src/main/java/com/kaua/template/domain/AggregateRoot.java
@@ -1,20 +1,0 @@
-package com.kaua.template.domain;
-
-import com.kaua.template.domain.events.DomainEvent;
-
-import java.util.List;
-
-public abstract class AggregateRoot<ID extends Identifier> extends Entity<ID> {
-
-    protected AggregateRoot(final ID id) {
-        super(id, 0);
-    }
-
-    protected AggregateRoot(final ID id, final long version) {
-        super(id, version);
-    }
-
-    protected AggregateRoot(final ID id, final long version, final List<DomainEvent> domainEvents) {
-        super(id, version, domainEvents);
-    }
-}
