--- domain/src/main/java/com/kaua/template/domain/Identifier.java
@@ -1,6 +1,0 @@
-package com.kaua.template.domain;
-
-public abstract class Identifier {
-
-    public abstract String getValue();
-}
