--- domain/src/main/java/com/kaua/template/domain/events/DomainEventPublisher.java
@@ -1,7 +1,0 @@
-package com.kaua.template.domain.events;
-
-@FunctionalInterface
-public interface DomainEventPublisher {
-
-    <T extends DomainEvent> void publish(T event);
-}
