--- domain/src/main/java/com/kaua/template/domain/exceptions/NoStackTraceException.java
@@ -1,12 +1,0 @@
-package com.kaua.template.domain.exceptions;
-
-public class NoStackTraceException extends RuntimeException {
-
-    public NoStackTraceException(final String message) {
-        this(message, null);
-    }
-
-    public NoStackTraceException(final String message, final Throwable cause) {
-        super(message, cause, true, false);
-    }
-}
