--- domain/src/main/java/com/kaua/template/domain/utils/IdUtils.java
@@ -1,20 +1,0 @@
-package com.kaua.template.domain.utils;
-
-import java.util.UUID;
-
-public final class IdUtils {
-
-    private IdUtils() {}
-
-    private static UUID generate() {
-        return UUID.randomUUID();
-    }
-    
-    public static String generateIdWithHyphen() {
-        return generate().toString();
-    }
-
-    public static String generateIdWithoutHyphen() {
-        return generate().toString().replace("-", "");
-    }
-}
