--- domain/src/main/java/com/kaua/template/domain/validation/Error.java
@@ -1,4 +1,0 @@
-package com.kaua.template.domain.validation;
-
-public record Error(String message) {
-}
