--- domain/src/main/java/com/kaua/template/domain/validation/Validator.java
@@ -1,16 +1,0 @@
-package com.kaua.template.domain.validation;
-
-public abstract class Validator {
-
-    private final ValidationHandler handler;
-
-    protected Validator(final ValidationHandler handler) {
-        this.handler = handler;
-    }
-
-    public abstract void validate();
-
-    protected ValidationHandler validationHandler() {
-        return this.handler;
-    }
-}
