+++ domain/src/test/java/com/kaua/order/domain/UnitTest.java
@@ -1,0 +1,7 @@
+package com.kaua.order.domain;
+
+import org.junit.jupiter.api.Tag;
+
+@Tag("unitTest")
+public class UnitTest {
+}
