--- domain/src/test/java/com/kaua/template/domain/UnitTest.java
@@ -1,7 +1,0 @@
-package com.kaua.template.domain;
-
-import org.junit.jupiter.api.Tag;
-
-@Tag("unitTest")
-public class UnitTest {
-}
