--- domain/src/test/java/com/kaua/template/domain/exceptions/NotFoundExceptionTest.java
@@ -1,47 +1,0 @@
-package com.kaua.template.domain.exceptions;
-
-import com.kaua.template.domain.AggregateRoot;
-import com.kaua.template.domain.Identifier;
-import com.kaua.template.domain.UnitTest;
-import com.kaua.template.domain.validation.ValidationHandler;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class NotFoundExceptionTest extends UnitTest {
-
-    @Test
-    void givenAValidAggregate_whenCallNotFoundExceptionWith_ThenReturnNotFoundException() {
-        // given
-        final var aggregate = SampleAggregate.class;
-        final var aId = "123";
-        final var expectedErrorMessage = "SampleAggregate with id 123 was not found";
-
-        // when
-        final var notFoundException = NotFoundException.with(aggregate, aId);
-        // then
-        Assertions.assertEquals(expectedErrorMessage, notFoundException.get().getMessage());
-    }
-
-    static class SampleAggregate extends AggregateRoot<SampleIdentifier> {
-        public SampleAggregate(SampleIdentifier id) {
-            super(id);
-        }
-
-        @Override
-        public void validate(ValidationHandler aHandler) {
-
-        }
-    }
-
-    static class SampleIdentifier extends Identifier {
-        private final String value;
-
-        public SampleIdentifier(String value) {
-            this.value = value;
-        }
-
-        public String getValue() {
-            return value;
-        }
-    }
-}
