--- domain/src/test/java/com/kaua/template/domain/utils/IdUtilsTest.java
@@ -1,25 +1,0 @@
-package com.kaua.template.domain.utils;
-
-import com.kaua.template.domain.UnitTest;
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-public class IdUtilsTest extends UnitTest {
-
-    @Test
-    void testCallIdUtilsGenerateIdWithHyphen() {
-        final var id = IdUtils.generateIdWithHyphen();
-
-        Assertions.assertNotNull(id);
-        Assertions.assertTrue(
-                id.matches("[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12}"));
-    }
-
-    @Test
-    void testCallIdUtilsGenerateIdWithoutHyphen() {
-        final var id = IdUtils.generateIdWithoutHyphen();
-
-        Assertions.assertNotNull(id);
-        Assertions.assertTrue(id.matches("[a-z0-9]{32}"));
-    }
-}
