+++ infrastructure/src/main/java/com/kaua/order/infrastructure/Main.java
@@ -1,0 +1,7 @@
+package com.kaua.order.infrastructure;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
