--- infrastructure/src/main/java/com/kaua/template/infrastructure/Main.java
@@ -1,7 +1,0 @@
-package com.kaua.template.infrastructure;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
